chức năng xóa xử lý, tìm kiếm sử ly, hàm kiểm tra dữ liệu thêm xử lý

+++ src/main/java/qltv/web/controllers/XuLyController.java
@@ -17,8 +17,10 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import qltv.web.dto.ThanhVienDTO;
 import qltv.web.dto.XuLyDTO;
+import qltv.web.mappers.ThanhVienMapper;
 import qltv.web.models.ThanhVien;
 import qltv.web.security.SecurityUtil;
 import qltv.web.services.ThanhVienService;
@@ -46,14 +48,14 @@ public class XuLyController {
         Long maTV = Long.parseLong(username);
         ThanhVienDTO user = thanhVienService.findMemberById(maTV);
         List<XuLyDTO> xuLys = xuLyService.getAllXuLy();
-        
+
         model.addAttribute("user", user);
         model.addAttribute("xuLys", xuLys);
         return "xu-ly-list";
     }
-    
+
     @GetMapping("/xuly/new")
-    public String CreateXuLy(Model model){
+    public String CreateXuLy(Model model) {
         String username = SecurityUtil.getUserSession();
         if (username == null) { // kiểm tra đăng nhập
             return "redirect:/login";
@@ -63,22 +65,22 @@ public class XuLyController {
         XuLyDTO xuLy = new XuLyDTO();
         Long nextMaXL = xuLyService.getMaxMaXL() + 1;
         List<ThanhVienDTO> thanhViens = thanhVienService.getAllThanhVien();
-        
+
         model.addAttribute("user", user);
         model.addAttribute("xuLy", xuLy);
         model.addAttribute("nextMaXL", nextMaXL);
         model.addAttribute("thanhViens", thanhViens);
-        
+
         return "xu-ly-create";
     }
-    
+
     @PostMapping("/xuly/new")
     public String saveXuLy(@RequestParam("maXuLy") String maXuLy,
-        @RequestParam("maThanhVien") String maThanhVien,
-     @RequestParam("hinhThucXuLy") String hinhThucXuLy,
-      @RequestParam("tienPhat") String tienPhat,
-         @RequestParam("trangThai") String trangThai,
-        Model model) {        
+            @RequestParam("maThanhVien") String maThanhVien,
+            @RequestParam("hinhThucXuLy") String hinhThucXuLy,
+            @RequestParam("tienPhat") String tienPhat,
+            @RequestParam("trangThai") String trangThai,
+            Model model) {
         String username = SecurityUtil.getUserSession();
         if (username == null) {
             return "redirect:/login";
@@ -90,7 +92,7 @@ public class XuLyController {
             return "xu-ly-create";
         }
 
-        switch(hinhThucXuLy){
+        switch (hinhThucXuLy) {
             case "KhoaThe2Thang":
                 hinhThucXuLy = "Khóa thẻ 2 tháng";
                 break;
@@ -107,22 +109,22 @@ public class XuLyController {
                 hinhThucXuLy = "Khóa thẻ 1 tháng";
         }
 
-        if(tienPhat.isEmpty()){
+        if (tienPhat.isEmpty()) {
             tienPhat = "0";
         }
 
         XuLyDTO xuLy = new XuLyDTO();
         xuLy.setMaXL(Integer.parseInt(maXuLy));
-        xuLy.setThanhVien(ThanhVienServiceImpl.mapToThanhVien(thanhVienService.findMemberById(Integer.parseInt(maThanhVien))));
+        xuLy.setThanhVien(thanhVienService.findMemberById(Integer.parseInt(maThanhVien)));
         xuLy.setHinhThucXL(hinhThucXuLy);
         xuLy.setSoTien(Integer.parseInt(tienPhat));
         xuLy.setTrangThaiXL(Integer.parseInt(trangThai));
         xuLy.setNgayXL(new Date());
-        
+
         xuLyService.saveXuLy(xuLy);
         return "redirect:/xuly";
     }
-    
+
     @GetMapping("/xuly/{maXL}/edit")
     public String editXuLy(@PathVariable("maXL") long maXL, Model model) {
         String username = SecurityUtil.getUserSession();
@@ -132,17 +134,17 @@ public class XuLyController {
         Long maTV = Long.parseLong(username);
         ThanhVienDTO user = thanhVienService.findMemberById(maTV);
         XuLyDTO xuLy = xuLyService.findXuLyByMaXL(maXL);
-        
+
         model.addAttribute("user", user);
         model.addAttribute("xuLy", xuLy);
-        
+
         return "xu-ly-edit";
     }
 
     @PostMapping("/xuly/{maXL}/edit")
     public String updateXuLy(@PathVariable("maXL") long maXL,
-                               @Valid @ModelAttribute("xuLy") XuLyDTO xuLy,
-                                BindingResult result, Model model) {
+            @Valid @ModelAttribute("xuLy") XuLyDTO xuLy,
+            BindingResult result, Model model) {
 
         String username = SecurityUtil.getUserSession();
         if (username == null) {
@@ -166,23 +168,23 @@ public class XuLyController {
         if (username == null) {
             return "redirect:/login";
         }
-        
-        xuLyService.deleteXuLy(maXL);
+
+        if (xuLyService.findXuLyByMaXL(maXL) != null) {
+            xuLyService.deleteXuLy(maXL);
+        }
         return "redirect:/xuly";
     }
 
     @GetMapping("/xuly/search")
-    public String searchThanhVien(@RequestParam(value = "query") String query, Model model) {
+    @ResponseBody
+    public List<XuLyDTO> searchThanhVien(@RequestParam(value = "query") String query, Model model) {
         String username = SecurityUtil.getUserSession();
         if (username == null) {
-            return "redirect:/login";
+            return null;
         }
         Long maTV = Long.parseLong(username);
         ThanhVienDTO user = thanhVienService.findMemberById(maTV);
         List<XuLyDTO> xuLys = xuLyService.searchXuLy(query);
-        
-        model.addAttribute("user", user);
-        model.addAttribute("xuLys", xuLys);
-        return "xu-ly-list";
+        return xuLys;
     }
 }
